Fix visualization-data P/Invoke type mismatch, document cache-drift edge case
29th /code-review high pass found two issues:

- cna_mediaplayer_get_visualization_data's count parameter was declared
  int, while its own doc comment claimed to match
  cna_vertexbuffer_get_data/cna_indexbuffer_get_data's established nuint
  length-parameter convention. A real mismatch: if native code declares
  the equivalent parameter as size_t, a 4-byte int argument leaves the
  upper register bytes undefined under the platform calling convention,
  which could inflate the count read far past the real 256-element
  buffers. Fixed by changing the parameter (and its call site) to nuint.
- IsVisualizationEnabled's C# cache could in principle desync from the
  real, native, audio-device-level state it mirrors (unlike Volume/
  IsMuted, which mirror pure C++ static state that can't change on its
  own). Checked whether the real engine's own GetVisualizationData has
  any signal to detect this before deciding not to fix it: it doesn't --
  "disabled" and "enabled but the device silently failed" both produce
  the same all-zero result there, by design, so a C# resync mechanism
  would have nothing more authoritative to check against. Documented
  rather than engineered around.

322/322 CNA.Framework.Tests + 91/91 CNA.XnaCompat.Tests passing,
unchanged. samples/HelloGame re-verified unaffected.

+++ src/CNA.Interop/Native.cs
@@ -429,8 +429,13 @@ internal static partial class Native
     /// bundled snapshot struct, matching <c>cna_vertexbuffer_set_data</c>/<c>get_data</c>'s own
     /// "explicit buffer, not a collection" convention for bulk binary data
     /// (<c>analysis_binding_sharp_runtime.md</c> §40) rather than inventing a new pattern for
-    /// exactly two fixed-size arrays.
+    /// exactly two fixed-size arrays. <paramref name="count"/> is <see cref="nuint"/>, matching
+    /// <c>cna_vertexbuffer_get_data</c>/<c>cna_indexbuffer_get_data</c>'s own length-parameter type
+    /// exactly (a code-review finding: an earlier version used <see langword="int"/> here despite
+    /// this doc comment's own claim to match that convention -- a real mismatch, not just cosmetic,
+    /// since a native side declaring the equivalent parameter as <c>size_t</c> would read
+    /// undefined upper bytes from a 4-byte argument under the platform calling convention).
     /// </summary>
     [LibraryImport(LibraryName)]
-    internal static unsafe partial CnaResult cna_mediaplayer_get_visualization_data(float* frequencies, float* samples, int count);
+    internal static unsafe partial CnaResult cna_mediaplayer_get_visualization_data(float* frequencies, float* samples, nuint count);
 }
